temporarily disable 00-tidy test
upstream depends on a perltidy version that is not yet in Debian, and
this test is mostly about preventing poorly formatted code being merged
upstream, so it is not a problem to disable it in Debian as long as one
checks the formatting of any patches sent upstream.

Remove this as soon as the required version is available in Debian.