enable grp geekotest to clone jobs
openqa-clone-job needs write access to .../factory/{iso,hdd} in order to
copy jobs to the local instance.  This ensures that these dirs belong to
grp 'geekotest', are writable and have the group S-bit set, so that one
can enable users to do this by adding them to the 'geekotest' group.

It also ensures that the group exists, and that the 'geekotest' user has
'geekotest' as its primary group.